+++ io-classes/src/Control/Monad/Class/MonadMVar.hs
@@ -1,3 +1,0 @@
-module Control.Monad.Class.MonadMVar {-# DEPRECATED "Use Control.Concurrent.Class.MonadMVar" #-} (module X) where
-
-import Control.Concurrent.Class.MonadMVar as X